Refactored commong login logic to LoggedInContainer
