+++ frontend/src/pages/BoardPage.tsx
@@ -1,0 +1,8 @@
+import BoardContainer from "@/components/board/BoardContainer"
+import LoggedInContainer from "@/components/LoggedInContainer"
+
+const ChartsContainer = () => {
+  return <LoggedInContainer>{(props) => <BoardContainer board={props.board} />}</LoggedInContainer>
+}
+
+export default ChartsContainer
